final_concur_scrubbing/requirements.txt
openai>=1.30.0
openpyxl>=3.1.2
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
prometheus-client>=0.20.0
structlog>=24.2.0
tenacity>=8.3.0
rich>=13.7.1
watchdog>=4.0.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
